WIP on refactor - has runtime module errors
