+++ ts/PlatformView.html
@@ -15,8 +15,9 @@
             href="https://fonts.googleapis.com/css?family=Nunito+Sans:200,600"
         />
         <link rel="stylesheet" type="text/css" href="./styles.css" />
+        <script>var exports = {"__esModule": true};</script>
         <script src="ptv-hmac-sha1.js"></script>
-        <script src="platformView.js"></script>
+        <script src="dist/platformView.js"></script>
     </head>
     <body onload="init()" data-colour="dark-blue">
         <div class="container">
